+++ minjun/07_DeepLearning/autoencoder_config.yaml
@@ -1,0 +1,13 @@
+seed: 17
+dataset_path: "/home/pervinco/Datasets/pytorch-mnist"
+save_path: "./runs"
+
+weight_init: True
+num_workers: 4
+
+num_labeled: 1
+num_unlabeled: 0
+
+epochs: 100
+batch_size: 64
+learning_rate: 0.001
